requirements.txt
fastapi
uvicorn
pandas
streamlit
requests
scikit-learn
google-generativeai


 
